+++ version/changelog.md
@@ -15,6 +15,7 @@ ou can find the binary downloads here: https://github.com/NHendricks/nh-toolbox/
   - initialize backup => init all values from previous repo
   - add number of files in repo in snapshot view
   - show number of files in snapshot
+  - restore single files
 
 ### v1.8.0 (10.02.2026)